Mostly score related fixes
When you fall too far, the game resets and sends you back to the main
menu.
The score you see on the HUD wont count down when you fall.
The most controls now require to be specified which controller to use.
The starting platforms are not weirdly random and you will always spawn
on one.